src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=-
#
argon2-cffi==21.1.0
    # via -r requirements.in
asgiref==3.4.1
    # via django
cffi==1.15.0
    # via argon2-cffi
django==3.2.9
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-debug-toolbar
    #   django-extensions
django-appconf==1.0.5
    # via django-compressor
django-compressor==2.4.1
    # via -r requirements.in
django-debug-toolbar==3.2.2
    # via -r requirements.in
django-environ==0.8.1
    # via django-environ-2
django-environ-2==2.3.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-nose==1.4.7
    # via -r requirements.in
nose==1.3.7
    # via django-nose
pycparser==2.21
    # via cffi
python3-memcached==1.51
    # via -r requirements.in
pytz==2021.3
    # via django
rcssmin==1.0.6
    # via django-compressor
rjsmin==1.1.0
    # via django-compressor
six==1.16.0
    # via django-compressor
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
whitenoise==5.3.0
    # via -r requirements.in
